+++ pkg/module/contact/service/svc_contact_email.go
@@ -54,6 +54,18 @@ func (svc *service) CreateContactEmail(ctx context.Context, contactId string, mo
 		return nil, contact.ErrContactNotFound
 	}
 
+	err = svc.validateContactEmail(ctx, parent, model)
+	if err != nil {
+		return nil, err
+	}
+
+	if model.IsDefault {
+		err = svc.resetDefaultContactEmail(ctx, parent, model)
+		if err != nil {
+			return nil, err
+		}
+	}
+
 	newId, err := svc.database.ContactEmailRepository().CreateContactEmail(ctx, parent, model)
 	if err != nil {
 		return nil, err
@@ -76,6 +88,22 @@ func (svc *service) UpdateContactEmail(ctx context.Context, contactId string, id
 	}
 	if data == nil {
 		return nil, contact.ErrContactEmailNotFound
+	}
+
+	data.Type = model.Type
+	data.Email = model.Email
+	data.IsDefault = model.IsDefault
+
+	err = svc.validateContactEmail(ctx, parent, data)
+	if err != nil {
+		return nil, err
+	}
+
+	if data.IsDefault {
+		err = svc.resetDefaultContactEmail(ctx, parent, data)
+		if err != nil {
+			return nil, err
+		}
 	}
 
 	err = svc.database.ContactEmailRepository().UpdateContactEmail(ctx, parent, data)
@@ -101,10 +129,40 @@ func (svc *service) DeleteContactEmail(ctx context.Context, contactId string, id
 	if data == nil {
 		return contact.ErrContactEmailNotFound
 	}
+	if data.IsDefault {
+		return contact.ErrContactEmailIsDefault
+	}
 
 	err = svc.database.ContactEmailRepository().DeleteContactEmail(ctx, parent, data)
 	if err != nil {
 		return err
 	}
+	return nil
+}
+
+func (svc *service) resetDefaultContactEmail(ctx context.Context, parent *model.Contact, model *model.Email) error {
+	current, err := svc.database.ContactEmailRepository().GetDefaultContactEmail(ctx, parent)
+	if err != nil {
+		return err
+	}
+	if current != nil && current.Id != model.Id {
+		current.IsDefault = false
+		err = svc.database.ContactEmailRepository().UpdateContactEmail(ctx, parent, current)
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func (svc *service) validateContactEmail(ctx context.Context, parent *model.Contact, model *model.Email) error {
+	modelType, err := svc.database.EmailTypeRepository().GetEmailTypeById(ctx, model.Type.Id)
+	if err != nil {
+		return err
+	}
+	if modelType == nil {
+		return contact.ErrEmailTypeNotFound
+	}
+
 	return nil
 }
